Extract canvas overlay components from PlotCanvas
