+++ docs/refactor-architecture-plan.md
@@ -288,11 +288,11 @@ Pull all extracted pieces together; simplify main PlotCanvas component:
 
 **Checklist 6:**
 
-- [ ] Import all utilities, components, and hooks
-- [ ] Consolidate PlotCanvas JSX to pure orchestration (render calls to: BedLayer, GridLayer, ToolpathOverlay, SelectionOverlay, HandleOverlay, GroupHandleOverlay, RulerOverlay, pen position crosshair)
+- [x] Import all utilities, components, and hooks (completed 2026-04-09)
+- [x] Consolidate PlotCanvas JSX to pure orchestration (render calls to: BedLayer, GridLayer, ToolpathOverlay, SelectionOverlay, HandleOverlay, GroupHandleOverlay, RulerOverlay, pen position crosshair) (completed 2026-04-09)
 - [ ] Remove all inline state, effects, event handlers (delegated to hooks)
-- [ ] Target size: PlotCanvas should drop from ~3100 → ~600–800 lines
-- [ ] Validate: `npm run typecheck` + `tests/component/PlotCanvas.test.tsx` (63/63) + `npm run test:e2e` (full suite)
+- [x] Target size: PlotCanvas should drop from ~3100 → ~600–800 lines (completed 2026-04-09: 614 lines)
+- [x] Validate: `npm run typecheck` + `tests/component/PlotCanvas.test.tsx` (63/63) + `npm run test:e2e` (full suite) (completed 2026-04-09)
 
 **Deliverable:** Step 6 complete when: PlotCanvas is primarily orchestration; all 1167+ unit tests + E2E tests pass; no regressions in canvas behavior.
 
@@ -561,6 +561,8 @@ Use this section to track completed steps with date and PR/commit references.
 - 2026-04-06: Expanded main-process coverage with dedicated tests for bootstrap, menu, events, and lightweight IPC registration modules (`tests/unit/main/bootstrap.*.test.ts`, `tests/unit/main/menu.applicationMenu.test.ts`, `tests/unit/main/events.pushEvents.test.ts`, `tests/unit/main/ipc.{app,fs,jobs,tasks}.test.ts`) plus broader branch coverage in config/fluidnc tests; `npx.cmd vitest run tests/unit/main` passes 37/37 and coverage is materially improved.
 - 2026-04-09: Started PlotCanvas Phase 5 keyboard extraction by adding `useCanvasKeyboardShortcuts` and moving inline key listeners out of `PlotCanvas.tsx`; updated `useSpaceKeyPan` to expose `setSpacePressed` while keeping pan gesture state local; validation passes with `npm.cmd run typecheck` and focused tests (`useCanvasKeyboardShortcuts`, `useSpaceKeyPan`, `PlotCanvas`) 82/82.
 - 2026-04-09: Completed Phase 5 validation with full E2E run passing 170/170 in 2.4 minutes; keyboard shortcut behavior remains stable through integration coverage.
+- 2026-04-09: Advanced Phase 6 by extracting in-file canvas overlays from `PlotCanvas.tsx` into `features/canvas/components/CanvasOverlays.tsx` and `features/canvas/components/CanvasSceneOverlays.tsx`; removed duplicated pen-crosshair render block; `PlotCanvas.tsx` reduced to 614 lines and validation remains green with `npm.cmd run typecheck` plus `tests/component/PlotCanvas.test.tsx` (63/63).
+- 2026-04-09: Completed Phase 6 validation pass with `npm.cmd run typecheck`, `tests/component/PlotCanvas.test.tsx` (63/63), and full `npm.cmd run test:e2e` run exiting cleanly (code 0).
 
 ## Update Rule
