+++ src/renderer/src/features/canvas/index.ts
@@ -10,4 +10,6 @@ export * from "./components/BedLayer";
 export * from "./components/GridLayer";
 export * from "./components/SelectionOverlay";
 export * from "./components/ToolpathOverlay";
+export * from "./components/CanvasOverlays";
+export * from "./components/CanvasSceneOverlays";
 export * from "./hooks";
